Refactoring: combined two apps together; Add: messages when add tracks
