+++ app/models.py
@@ -1,5 +1,17 @@
+from django.contrib.auth.models import AbstractUser
 from django.db import models
-from accounts.models import CustomUser
+from django.utils.translation import gettext_lazy as _
+
+
+
+class CustomUser(AbstractUser):
+    email = models.EmailField(
+        _("email address"),
+        blank=False,
+        unique=True
+        )
+    added_tracks = models.ManyToManyField('app.Tracks', through='app.UserHasTracks')
+    added_playlists = models.ManyToManyField('app.Playlists', through='app.UserHasPlaylists')
 
 
 class Albums(models.Model):
@@ -63,7 +75,7 @@ class Playlists(models.Model):
     
     def __repr__(self):
         return "Playlist(name={}, creator={})".format(self.name, self.creator)
-
+    
 
 class Comments(models.Model):
     message = models.TextField()
