+++ app/templates/base.html
@@ -13,8 +13,8 @@
     <section class="main">
         <div class="sidebar">
             <ul>
-                <a href="{% url 'top-artists' %}">Лучшие исполнители</a>
-                <a href="{% url 'top-tracks' %}">Горячие треки</a>
+                <a href="{% url 'top_artists' %}">Лучшие исполнители</a>
+                <a href="{% url 'top_tracks' %}">Горячие треки</a>
                 {% if request.user.is_authenticated %}
                 <a href="{% url 'create_playlist' %}" >Cоздать плейлист</a>
                 {% endif %}
