+++ app/templates/content/playlist.html
@@ -7,7 +7,9 @@
 
 <h2>Треки</h2>
 {% for track in tracks %}
-<p>{{ track.name }}</p>
-<p>{{ track.author }}</p>
+<p>
+    Трек "{{ track.name }}" ({{ track.author }})
+</p>
+
 {% endfor %}
 {% endblock %}
