+++ app/templates/content/searching.html
@@ -14,9 +14,11 @@
           Исполнитель: {{ track.track.artist.name }}
         </p>
         {% if request.user.is_authenticated%}
-          <form method="POST" action="{% url 'top-tracks' %}?next={{ request.path }}&query={{ query }}">
+          <form method="post" action="{% url 'add_track' %}">
             {% csrf_token %}
             <input type="hidden" name="pk" value="{{ track.track.id_musixmatch }}">
+            <input type="hidden" name="next" value="{{ request.path }}">
+            <input type="hidden" name="query" value="{{ query }}">
             <input type="submit" value="Добавить">
           </form>
         {% endif %}
