+++ app/templates/content/top_tracks.html
@@ -11,9 +11,10 @@
         Исполнитель: {{ track.track.artist.name }}
       </p>
       {% if request.user.is_authenticated %}
-      <form method="POST" action="{% url 'top-tracks' %}">
+      <form method="POST" action="{% url 'add_track' %}">
         {% csrf_token %}
         <input type="hidden" name="pk" value="{{ track.track.id_musixmatch }}">
+        <input type="hidden" name="next" value="{{ request.path }}">
         <input type="submit" value="Добавить">
       </form>
       {% endif %}
