+++ app/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import path, include
 
 
 from . import views
@@ -6,8 +6,14 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-    path('top-artists', views.TopArtistsView.as_view(), name='top-artists'),
-    path('top-tracks', views.TopTracksView.as_view(), name='top-tracks'),
+    path('login/', views.MyLoginView.as_view(), name='login'),
+    path('', include('django.contrib.auth.urls')),
+    path('add_track', views.add_track, name='add_track'),
+    path('create-playlist/', views.CreatePlaylistView.as_view(), name='create_playlist'), 
+    path('playlist/<int:pk>', views.playlist, name='playlist'),
+    path('profile/', views.profile, name='profile'),
+    path('register/', views.RegistrationView.as_view(), name='registration'),
     path('search', views.SearchView.as_view(), name='search'),
-    path('playlist/<int:pk>', views.playlist, name='playlist')
+    path('top-artists', views.TopArtistsView.as_view(), name='top_artists'),
+    path('top-tracks', views.TopTracksView.as_view(), name='top_tracks'),    
 ]
